requirements.txt
apache-airflow==2.10.5
kafka-python==2.1.4
requests==2.31.0
datetime== 5.3
jsonschema==4.19.0